refactor: move proxy entrypoints component in api package
