+++ gravitee-apim-console-webui/src/management/api/apis-routing.module.ts
@@ -32,7 +32,7 @@ import { ApiGeneralMembersComponent } from './general/user-group-access/members/
 import { ApiGeneralGroupsComponent } from './general/user-group-access/groups/api-general-groups.component';
 import { ApiGeneralTransferOwnershipComponent } from './general/user-group-access/transfer-ownership/api-general-transfer-ownership.component';
 import { ApiPortalDocumentationMetadataComponent } from './general/documentation/metadata/api-portal-documentation-metadata.component';
-import { ApiProxyEntrypointsComponent } from './proxy/entrypoints/api-proxy-entrypoints.component';
+import { ApiEntrypointsComponent } from './entrypoints/api-entrypoints.component';
 import { ApiCorsComponent } from './cors/api-cors.component';
 import { ApiProxyResponseTemplatesListComponent } from './proxy/response-templates/list/api-proxy-response-templates-list.component';
 import { ApiProxyResponseTemplatesEditComponent } from './proxy/response-templates/edit/api-proxy-response-templates-edit.component';
@@ -535,7 +535,7 @@ const apisRoutes: Routes = [
        */
       {
         path: 'v2/entrypoints',
-        component: ApiProxyEntrypointsComponent,
+        component: ApiEntrypointsComponent,
         data: {
           apiPermissions: {
             only: ['api-definition-r', 'api-health-r'],
