+++ gravitee-apim-console-webui/src/management/api/entrypoints/api-entrypoints.component.html
