+++ gravitee-apim-console-webui/src/management/api/entrypoints/api-entrypoints.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../scss/gio-layout' as gio-layout;
+@use '../../../scss/gio-layout' as gio-layout;
 
 :host {
   @include gio-layout.gio-responsive-content-container;
